Remove conversion warning from src/geomgraph directory 	modified:   src/geomgraph/Node.cpp 	modified:   src/geomgraph/index/MonotoneChainEdge.cpp 	modified:   src/geomgraph/index/SimpleEdgeSetIntersector.cpp 	modified:   src/geomgraph/index/SimpleMCSweepLineIntersector.cpp 	modified:   src/geomgraph/index/SimpleSweepLineIntersector.cpp

+++ src/geomgraph/Node.cpp
@@ -261,7 +261,7 @@ Node::addZ(double z)
 	}
 	zvals.push_back(z);
 	ztot+=z;
-	coord.z=ztot/zvals.size();
+	coord.z = ztot / static_cast<double>(zvals.size());
 #if GEOS_DEBUG
 	cerr<<" added "<<z<<": ["<<ztot<<"/"<<zvals.size()<<"="<<coord.z<<"]"<<endl;
 #endif
